refactor(sip): extract the Replaces dialog terminator (R3)
Adding the empty-user-part branch to the From URI pushed
SipCallSignalingService past the 1000-line cap: 999 before, 1003 after, and the
architecture baseline for that rule is empty, so the gate went red with
"beheben, nicht baselinen".

TerminateReplacedDialogAsync moves out as SipReplacedDialogTerminator — a real
extract class, not a partial split. The session lookup stays at the call site, so
the collaborator needs no view of the service's session table: telemetry and
logger are all it takes.

Worth noting for whoever touches this file next: SipCallSignalingHelpers already
says in its own docs that it was extracted "to keep the signaling service within
the per-file size limit". Those 999 lines were the remainder of an earlier
rescue, not headroom.

969 lines now. ArchitectureTests 14/14, 8358 unit tests green, 71 of 74 interop
green with zero transfer failures — the Replaces path runs straight through the
extracted code.

Refs #104
